% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayr-package.R
\docType{package}
\name{bayr-package}
\alias{bayr}
\alias{bayr-package}
\title{bayr: Tidy and Unified Reporting of Bayesian Regression Models}
\description{
Provides a unified, tidy interface for reporting Bayesian regression models
in the spirit of the Bayesian New Statistics. Wrapper functions extract
parameter estimates and posterior distributions from 'brms' and 'rstanarm'
models, and tidy summaries from 'lme4' fits, into a common long format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://schmettow.github.io/New_Stats/}
  \item \url{https://github.com/schmettow/bayr}
  \item Report bugs at \url{https://github.com/schmettow/bayr/issues}
}

}
\author{
\strong{Maintainer}: Martin Schmettow \email{schmettow@web.de}

}
\keyword{internal}
