% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_extraction.R, R/markup_helpers.R
\name{coef.tbl_post}
\alias{coef.tbl_post}
\alias{coef.data.frame}
\alias{coef.MCMCglmm}
\alias{coef.brmsfit}
\alias{coef.stanfit}
\alias{coef.stanreg}
\alias{fixef}
\alias{fixef.tbl_post}
\alias{fixef.MCMCglmm}
\alias{fixef.brmsfit}
\alias{fixef.stanfit}
\alias{fixef.stanreg}
\alias{ranef}
\alias{ranef.tbl_post}
\alias{ranef.MCMCglmm}
\alias{ranef.brmsfit}
\alias{ranef.stanfit}
\alias{ranef.stanreg}
\alias{grpef}
\alias{grpef.tbl_post}
\alias{grpef.MCMCglmm}
\alias{grpef.brmsfit}
\alias{grpef.stanfit}
\alias{grpef.stanreg}
\alias{print.tbl_coef}
\alias{knit_print.tbl_coef}
\title{Coefficient extraction}
\usage{
\method{coef}{tbl_post}(
  object,
  model = unique(object$model),
  type = c("fixef", "ranef"),
  mean.func = identity,
  estimate = median,
  interval = 0.95,
  ...
)

\method{coef}{data.frame}(object, ...)

\method{coef}{MCMCglmm}(object, estimate = median, ...)

\method{coef}{brmsfit}(object, estimate = median, ...)

\method{coef}{stanfit}(object, estimate = median, ...)

\method{coef}{stanreg}(object, estimate = median, ...)

fixef(object, estimate = median, ...)

\method{fixef}{tbl_post}(object, estimate = median, ...)

\method{fixef}{MCMCglmm}(object, estimate = median, ...)

\method{fixef}{brmsfit}(object, estimate = median, ...)

\method{fixef}{stanfit}(object, estimate = median, ...)

\method{fixef}{stanreg}(object, estimate = median, ...)

ranef(object, estimate = median, ...)

\method{ranef}{tbl_post}(object, estimate = median, ...)

\method{ranef}{MCMCglmm}(object, estimate = median, ...)

\method{ranef}{brmsfit}(object, estimate = median, ...)

\method{ranef}{stanfit}(object, estimate = median, ...)

\method{ranef}{stanreg}(object, estimate = median, ...)

grpef(object, estimate = median, ...)

\method{grpef}{tbl_post}(object, estimate = median, ...)

\method{grpef}{MCMCglmm}(object, estimate = median, ...)

\method{grpef}{brmsfit}(object, estimate = median, ...)

\method{grpef}{stanfit}(object, estimate = median, ...)

\method{grpef}{stanreg}(object, estimate = median, ...)

\method{print}{tbl_coef}(x, ...)

\method{knit_print}{tbl_coef}(x, ...)
}
\arguments{
\item{object}{tbl_post (brms, MCMCglmm) object holding the posterior in long format}

\item{model}{model}

\item{type}{type of coefficient: fixef (grpef, ranef)}

\item{mean.func}{function (identity)}

\item{estimate}{function for computing the center estimate (posterior mode)}

\item{interval}{credibility interval: .95}

\item{...}{ignored}

\item{x}{tbl_coef object}
}
\value{
coefficient table with parameter name, estimate and interval

The standard center function is the posterior median
}
\description{
summary table of fixed, random or group-level coefficients and fitted values (eta,
only stanfit models) from posterior
}
\author{
Martin Schmettow
}
