% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_extensions.R
\name{discard_all_na}
\alias{discard_all_na}
\title{Removing completely unmeasured variables}
\usage{
discard_all_na(D)
}
\arguments{
\item{D}{data frame}
}
\value{
data frame
}
\description{
all columns that are completely NA are discarded
}
\author{
Martin Schmettow
}
