% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_extensions.R
\name{discard_redundant.tbl_clu}
\alias{discard_redundant.tbl_clu}
\alias{discard_redundant.tbl_coef}
\alias{discard_redundant.tbl_post_pred}
\alias{discard_redundant.tbl_predicted}
\alias{discard_redundant.tbl_post}
\title{Removing redundant variables}
\usage{
discard_redundant.tbl_clu(object, except = c())

discard_redundant.tbl_coef(object, except = c())

discard_redundant.tbl_post_pred(object, except = c())

discard_redundant.tbl_predicted(object, except = c())

discard_redundant.tbl_post(object, except = c())
}
\arguments{
\item{except}{vector of column names to keep}

\item{D}{data frame}
}
\value{
data frame
}
\description{
all variables that do not vary are discarded
}
\author{
Martin Schmettow
}
