% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_extensions.R
\name{expand_grid_df}
\alias{expand_grid_df}
\title{Tibble-returning expand.grid}
\usage{
expand_grid_df(...)
}
\arguments{
\item{...}{factors}
}
\value{
tibble
}
\description{
works like expand.grid, but returns a tibble. Named `expand_grid_df()` to
avoid masking `tidyr::expand_grid()`.
}
\author{
Martin Schmettow
}
