% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{formula.MCMCglmm}
\alias{formula.MCMCglmm}
\title{formula interface for MCMCglmm}
\usage{
formula()
}
\arguments{
\item{model}{MCMCglmm object

only fixed part implemented by now}
}
\description{
returns the formula of an MCMCglmm object
}
\author{
Martin Schmettow
}
