% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_extensions.R
\name{go_first}
\alias{go_first}
\alias{go_arrange}
\title{Moving column to first position}
\usage{
go_first(D, ...)

go_arrange(D, ...)
}
\arguments{
\item{D}{data frame}

\item{...}{columns to move, using dplyr::select tidyselect semantics}
}
\value{
data frame

go_first and go_arrange expect a column specification, similar to
dplyr::select
}
\description{
The referred column becomes the first, keeping the order of the others
}
\examples{
D <- tibble::tibble(x = 1:3, y = 6:4, z = c(8,9,7))
go_first(D, y)
go_first(D, y:z)
go_arrange(D, y)



}
\author{
Martin Schmettow
}
