% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_extraction.R
\name{join.tbl_coef}
\alias{join.tbl_coef}
\title{Joining two coefficent tables (NI)}
\usage{
join.tbl_coef(x, y)
}
\arguments{
\item{first}{coefficient table tbl_coef or tbl_compcoef}

\item{second}{coefficient table tbl_coef}

\item{modelnames}{list of model names}
}
\value{
data_frame with parameter names and point estimates

Returns a comparative coefficient table (tbl_coefcomp).
Models are shown in columns, Parameters ordered by the most complex model
Models ordered by number of parameters (nesting?)
Model names are taken via lazy eval, or can be given.
}
\description{
NOT IMPLEMENTED
creates an overview table of point estimates for models with overlapping
parameter sets
}
\author{
Martin Schmettow
}
