% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_helpers.R
\name{md_coef}
\alias{md_coef}
\alias{frm_coef}
\title{coefficient formula}
\usage{
md_coef(
  tbl_coef,
  ...,
  row = NULL,
  center = T,
  interval = F,
  prefix = F,
  round = 2,
  mean_fnc = identity,
  neg = F
)

frm_coef(
  tbl_coef,
  ...,
  row = NULL,
  center = T,
  interval = T,
  prefix = F,
  round = 2,
  neg = F
)
}
\arguments{
\item{tbl_coef}{coefficient table}

\item{...}{filter formulas, using annotations}

\item{row}{row number of parameter}

\item{center}{add center estimate (TRUE)}

\item{interval}{add interval (TRUE)}

\item{prefix}{add prefix term with coef name (not implemented)}

\item{round}{digits (2)}

\item{mean_fnc}{mean function (identity)}

\item{neg}{output negative estimate}
}
\value{
markdown string

The parameter can alternatively selected by row number, or using a
dplyr-like selection formula. When the selection criteria are ambiguous
an error is thrown
}
\description{
inline markup reporting of estimates with credibility limits
}
\author{
Martin Schmettow
}
