% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_helpers.R
\name{mrk_coef}
\alias{mrk_coef}
\title{mrk_coef}
\usage{
mrk_coef(tbl_coef, coef, center = T, interval = T, prefix = F,
  round = 2)
}
\arguments{
\item{tbl_coef}{coefficient table}

\item{coef}{coefficient (row number or name)}

\item{center}{add center estimate (TRUE)}

\item{interval}{add interval (TRUE)}

\item{prefix}{add prefix term with coef name (not implemented)}

\item{rounding}{digits (2)}
}
\value{
markdown string
}
\description{
inline markup reporting of estimates with credibility limits (formula)
}
\author{
Martin Schmettow
}

