% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_helpers.R, R/postpred_extraction.R
\name{print.tbl_post_pred}
\alias{print.tbl_post_pred}
\alias{print.tbl_predicted}
\alias{knit_print.tbl_post_pred}
\alias{knit_print.tbl_predicted}
\alias{post_pred}
\alias{mtx_post_pred}
\alias{mtx_post_pred.brmsfit}
\alias{mtx_post_pred.stanreg}
\title{posterior predictive extraction}
\usage{
\method{print}{tbl_post_pred}(x, ...)

\method{print}{tbl_predicted}(x, ...)

\method{knit_print}{tbl_post_pred}(x, ...)

\method{knit_print}{tbl_predicted}(x, ...)

post_pred(
  model,
  scale = "resp",
  model_name = deparse(substitute(model)),
  newdata = NULL,
  thin = 1,
  ...
)

mtx_post_pred(model, ...)

\method{mtx_post_pred}{brmsfit}(model, model_name, newdata = NULL, thin = 1, ...)

\method{mtx_post_pred}{stanreg}(model, model_name, newdata = NULL, thin = 1, ...)
}
\arguments{
\item{x}{tbl_post_pred object}

\item{...}{ignored}

\item{model}{Bayesian model object}

\item{scale}{"response" or "lin_pred"}

\item{model_name}{provides a name for the model}

\item{newdata}{new data to predict from}

\item{thin}{thinning factor}
}
\value{
tbl_postpred object with MCMC draws

chains are stored in a
long table with the following columns:

chain  iter   Obs   value  type  order
(fctr) (int)  (int)       (dbl)  (chr) (int)
}
\description{
MCMC predicted values are extracted from a Bayesian (regression) object
and returned as a tidy tbl_post_pred object.
}
\author{
Martin Schmettow
}
