% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_extraction.R
\name{predict.tbl_post_pred}
\alias{predict.tbl_post_pred}
\alias{predict.brmsfit}
\alias{predict.stanreg}
\title{Posterior predictions}
\usage{
\method{predict}{tbl_post_pred}(object, scale = c("resp"), center = median, interval = 0.95, ...)

\method{predict}{brmsfit}(object, scale = c("resp"), center = median, interval = 0.95, ...)

\method{predict}{stanreg}(object, scale = c("resp"), center = median, interval = 0.95, ...)
}
\arguments{
\item{object}{regression model or tbl_postpred}

\item{scale}{linpred or resp}

\item{center}{function for computing the center estimate (median)}

\item{interval}{credibility interval: .95}

\item{...}{passing parameters to tbl_postpred (e.g. thin)}
}
\value{
coefficient table with center and interval estimates per obs

The standard center function is the posterior median
}
\description{
summary table of predicted values from predictive posterior
}
\author{
Martin Schmettow
}
