% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_extraction.R
\name{seperate.tbl_coef}
\alias{seperate.tbl_coef}
\title{separating factors an group means model}
\usage{
seperate.tbl_coef(x, y)
}
\arguments{
\item{first}{coefficient table tbl_coef or tbl_compcoef}

\item{second}{coefficient table tbl_coef}

\item{modelnames}{list of model names}
}
\value{
data_frame with two or more factor variables, center and CI.

Returns a data_frame with parameter plit into factors and cleaned.
Prepares for intercation plots
}
\description{
NOT IMPLEMENTED
In a model with interaction effects only (group means contrasts)
coefficient names are split into factor variables (and cleaned)
}
\author{
Martin Schmettow
}
