% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_extraction.R
\name{tbl_post}
\alias{posterior}
\alias{tbl_post}
\alias{tbl_post.MCMCglmm}
\alias{tbl_post.brmsfit}
\title{posterior extraction (c}
\usage{
tbl_post(model, ...)

posterior(model, ...)

\method{tbl_post}{MCMCglmm}(model, ...)

\method{tbl_post}{brmsfit}(model, ...)
}
\arguments{
\item{model}{Bayesian model object}

\item{...}{ignored}
}
\value{
tbl_post object with MCMC chain in long format

The MCMC chains are extracted from the model and stored in a
common format. Different to most internal representations, and
mcmc objects from package coda, the chains are stored in a
long table with the following columns:
}
\description{
MCMC chains are  extracted from a Bayesian (regression) object
and returned as a posterior object, which is in long format
(chain, iter, parameter, value, type, order). Parameters are classified
as fixef, ranef, grpef and named after a common scheme.
}
\author{
Martin Schmettow
}

