% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_extensions.R
\name{update_by}
\alias{update_by}
\title{update columns conditionally}
\usage{
update_by(D, by, ...)
}
\arguments{
\item{D}{data frame}

\item{by}{predicate function (like dplyr::filter)}

\item{...}{expressions}
}
\value{
data frame

Applies mutations to the filtered group, only.
}
\description{
update columns conditionally
}
\examples{
D <- tibble::tibble(group = c(1, 1, 2, 2), value = c(4, 9, -4, -9))

dplyr::mutate(D, value = dplyr::if_else(group == 1, sqrt(value), value))
## Produces NaNs, because sqrt() is evaluated before selection

update_by(D, group == 1, value = sqrt(value))
## sqrt() is only evaluated

}
\author{
Martin Schmettow
}
