% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_extensions.R
\name{varies}
\alias{varies}
\title{Varies}
\usage{
varies(x)
}
\arguments{
\item{D}{data frame}
}
\value{
data frame
}
\description{
tests if a variable has more than one level.
}
\details{
TODO:
+ transfer to remodels.
+ think about NA handling
}
\author{
Martin Schmettow
}
