% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_extraction.R, R/markup_helpers.R
\name{clu}
\alias{clu}
\alias{assert_clu}
\alias{clu.tbl_post}
\alias{clu.data.frame}
\alias{clu.MCMCglmm}
\alias{clu.brmsfit}
\alias{clu.stanfit}
\alias{clu.stanreg}
\alias{clu.glmerMod}
\alias{print.tbl_clu}
\alias{knit_print.tbl_clu}
\title{center-lower-upper table}
\usage{
clu(object, ...)

assert_clu(x)

\method{clu}{tbl_post}(
  object,
  model = unique(object$model),
  mean.func = identity,
  estimate = median,
  interval = 0.95,
  ...
)

\method{clu}{data.frame}(object, ...)

\method{clu}{MCMCglmm}(object, ...)

\method{clu}{brmsfit}(object, ...)

\method{clu}{stanfit}(object, ...)

\method{clu}{stanreg}(object, ...)

\method{clu}{glmerMod}(
  object,
  mean.func = identity,
  model_name = as.character(deparse(substitute(object))),
  interval = 0.95,
  ...
)

\method{print}{tbl_clu}(x, ...)

\method{knit_print}{tbl_clu}(x, ...)
}
\arguments{
\item{object}{tbl_post (brms, MCMCglmm) object holding the posterior in long format}

\item{...}{ignored}

\item{x}{a tbl_clu object}

\item{model}{name of model}

\item{mean.func}{function (identity)}

\item{estimate}{function for computing the center estimate (posterior mode)}

\item{interval}{credibility interval: .95}

\item{model_name}{name of the model}
}
\value{
CLU table (tbl_clu) with parameter name, estimate and interval.
}
\description{
returns a summary table for estimates with median (center) and 95% limits'
}
\author{
Martin Schmettow
}
