% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_extraction.R
\name{coef.tbl_post}
\alias{coef.tbl_post}
\alias{fixef.MCMCglmm}
\alias{fixef.brmsfit}
\alias{fixef.tbl_post}
\alias{grpef}
\alias{grpef.MCMCglmm}
\alias{grpef.brmsfit}
\alias{grpef.tbl_post}
\alias{ranef.MCMCglmm}
\alias{ranef.brmsfit}
\alias{ranef.tbl_post}
\title{Coefficient extraction}
\usage{
\method{coef}{tbl_post}(object, type = "fixef", estimate = shorth,
  interval = 0.95, ...)

\method{fixef}{tbl_post}(object, estimate = shorth, ...)

\method{fixef}{MCMCglmm}(object, estimate = shorth, ...)

\method{fixef}{brmsfit}(object, estimate = shorth, ...)

\method{ranef}{tbl_post}(object, estimate = shorth, ...)

\method{ranef}{MCMCglmm}(object, estimate = shorth, ...)

\method{ranef}{brmsfit}(object, estimate = shorth, ...)

grpef(object, estimate = shorth, ...)

\method{grpef}{tbl_post}(object, estimate = shorth, ...)

\method{grpef}{MCMCglmm}(object, estimate = shorth, ...)

\method{grpef}{brmsfit}(object, estimate = shorth, ...)
}
\arguments{
\item{object}{tbl_post (brms, MCMCglmm) object holding the posterior in long format}

\item{type}{type of coefficient: fixef (grpef, ranef)}

\item{estimate}{function for computing the center estimate (posterior mode)}

\item{interval}{credibility interval: .95}

\item{...}{ignored}
}
\value{
coefficient table with parameter name, estimate and interval

The standard center function is the posterior mode computed
by modeest::shorth
}
\description{
summary table of fixed, random or group-level coefficients from posterior
}
\author{
Martin Schmettow
}

