% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_extraction.R
\name{coef.tbl_post}
\alias{coef.MCMCglmm}
\alias{coef.brms}
\alias{coef.stanfit}
\alias{coef.stanreg}
\alias{coef.tbl_post}
\alias{fitted.brmsfit}
\alias{fitted.tbl_post}
\alias{fixef.MCMCglmm}
\alias{fixef.brmsfit}
\alias{fixef.stanfit}
\alias{fixef.stanreg}
\alias{fixef.tbl_post}
\alias{grpef}
\alias{grpef.MCMCglmm}
\alias{grpef.brmsfit}
\alias{grpef.stanfit}
\alias{grpef.stanreg}
\alias{grpef.tbl_post}
\alias{print.tbl_coef}
\alias{ranef.MCMCglmm}
\alias{ranef.brmsfit}
\alias{ranef.stanfit}
\alias{ranef.stanreg}
\alias{ranef.tbl_post}
\title{Coefficient extraction}
\usage{
\method{coef}{tbl_post}(object, model = unique(object$model),
  type = c("fixef", "disp", "grpef"), mean.func = identity,
  estimate = shorth, interval = 0.95, ...)

\method{print}{tbl_coef}(x, digits = NULL, title = T, footnote = T,
  kable = by_knitr(), ...)

\method{coef}{MCMCglmm}(object, estimate = shorth, ...)

\method{coef}{brms}(object, estimate = shorth, ...)

\method{coef}{stanfit}(object, estimate = shorth, ...)

\method{coef}{stanreg}(object, estimate = shorth, ...)

\method{fixef}{tbl_post}(object, estimate = shorth, ...)

\method{fixef}{MCMCglmm}(object, estimate = shorth, ...)

\method{fixef}{brmsfit}(object, estimate = shorth, ...)

\method{fixef}{stanfit}(object, estimate = shorth, ...)

\method{fixef}{stanreg}(object, estimate = shorth, ...)

\method{ranef}{tbl_post}(object, estimate = shorth, ...)

\method{ranef}{MCMCglmm}(object, estimate = shorth, ...)

\method{ranef}{brmsfit}(object, estimate = shorth, ...)

\method{ranef}{stanfit}(object, estimate = shorth, ...)

\method{ranef}{stanreg}(object, estimate = shorth, ...)

grpef(object, estimate = shorth, ...)

\method{grpef}{tbl_post}(object, estimate = shorth, ...)

\method{grpef}{MCMCglmm}(object, estimate = shorth, ...)

\method{grpef}{brmsfit}(object, estimate = shorth, ...)

\method{grpef}{stanfit}(object, estimate = shorth, ...)

\method{grpef}{stanreg}(object, estimate = shorth, ...)

\method{fitted}{tbl_post}(object, estimate = shorth, ...)

\method{fitted}{brmsfit}(object, estimate = shorth, ...)
}
\arguments{
\item{object}{tbl_post (brms, MCMCglmm) object holding the posterior in long format}

\item{model}{model}

\item{type}{type of coefficient: fixef (grpef, ranef)}

\item{estimate}{function for computing the center estimate (posterior mode)}

\item{interval}{credibility interval: .95}

\item{...}{ignored}

\item{mean}{function (identity)}
}
\value{
coefficient table with parameter name, estimate and interval

The standard center function is the posterior mode computed
by modeest::shorth
}
\description{
summary table of fixed, random or group-level coefficients and fitted values (eta,
only stanfit models) from posterior

fitted <-
	function(object, estimate = shorth, ...) UseMethod("fitted", object)
}
\author{
Martin Schmettow
}

