% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{compare_IC}
\alias{compare_IC}
\title{Information Criteria comparison}
\usage{
compare_IC(ic_list)
}
\arguments{
\item{ic}{loo.psis, kfold or waic object}
}
\value{
tbl_df
}
\description{
Collects results from multiple Loo objects and
puts them into a tidy model comparison table.
}
\author{
Martin Schmettow
}
