% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_helpers.R, R/model_selection.R
\name{print.tbl_IC_comp}
\alias{print.tbl_IC_comp}
\alias{knit_print.tbl_IC_comp}
\alias{compare_IC}
\title{Information Criteria comparison}
\usage{
\method{print}{tbl_IC_comp}(x, ...)

knit_print.tbl_IC_comp(x, ...)

compare_IC(ic_list)
}
\arguments{
\item{ic}{loo.psis, kfold or waic object}
}
\value{
tbl_df
}
\description{
Collects results from multiple Loo objects and
puts them into a tidy model comparison table.
}
\author{
Martin Schmettow
}
