% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_extraction.R
\name{fixef_ml}
\alias{fixef_ml}
\title{Multi-level coefficient table}
\usage{
fixef_ml(object, ...)
}
\arguments{
\item{object}{tbl_post (brms, rstanarm) object holding the posterior in long format}

\item{...}{passed on to grpef and fixef}

\item{model}{model}
}
\value{
coefficient table with standard deviations
}
\description{
produces a CLU table of population-level effects together with
random effects standard deviation for all random factor levels.
}
\author{
Martin Schmettow
}
