% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{IC}
\alias{IC}
\title{Information Criteria extraction}
\usage{
IC(ic)
}
\arguments{
\item{ic}{loo.psis, kfold or waic object}
}
\value{
tbl_df
}
\description{
Tidies the output of model evaluation commands from package Loo.
}
\author{
Martin Schmettow
}
