% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_helpers.R, R/model_selection.R
\name{print.tbl_IC}
\alias{print.tbl_IC}
\alias{knit_print.tbl_IC}
\alias{IC}
\title{Information Criteria extraction}
\usage{
\method{print}{tbl_IC}(x, ...)

\method{knit_print}{tbl_IC}(x, ...)

IC(ic)
}
\arguments{
\item{ic}{loo.psis, kfold or waic object}
}
\value{
tbl_df
}
\description{
Tidies the output of model evaluation commands from package Loo.
}
\author{
Martin Schmettow
}
