% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpred_extraction.R
\name{post_pred}
\alias{post_pred}
\alias{tbl_post_pred}
\alias{print.tbl_post_pred}
\alias{tbl_post_pred.data.frame}
\alias{tbl_post_pred.brmsfit}
\alias{tbl_post_pred.stanreg}
\title{posterior predictive extraction}
\usage{
post_pred(model, scale = "obs", model_name, thin)

tbl_post_pred(model, ...)

\method{print}{tbl_post_pred}(tbl_post_pred, kable = by_knitr(), ...)

\method{tbl_post_pred}{data.frame}(df, ...)

\method{tbl_post_pred}{brmsfit}(model, newdata = NULL, ...)

\method{tbl_post_pred}{stanreg}(model, newdata = NULL, ...)
}
\arguments{
\item{model}{Bayesian model object}

\item{scale}{"response" or "lin_pred"}

\item{model_name}{provides a name for the model}

\item{thin}{thinning factor}

\item{newdata}{new data to predict from}
}
\value{
tbl_postpred object with MCMC draws

chains are stored in a
long table with the following columns:

chain  iter   parameter   value  type  order
(fctr) (int)  (chr)       (dbl)  (chr) (int)
}
\description{
MCMC predicted values are  extracted from a Bayesian (regression) object
and returned as a tbl_post_pred object
}
\author{
Martin Schmettow
}
