% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_helpers.R, R/postpred_extraction.R
\name{print.tbl_post_pred}
\alias{print.tbl_post_pred}
\alias{print.tbl_predicted}
\alias{knit_print.tbl_post_pred}
\alias{knit_print.tbl_predicted}
\alias{post_pred}
\title{posterior predictive extraction}
\usage{
\method{print}{tbl_post_pred}(x, ...)

\method{print}{tbl_predicted}(x, ...)

\method{knit_print}{tbl_post_pred}(x, ...)

\method{knit_print}{tbl_predicted}(x, ...)

post_pred(model, scale = "obs", model_name, thin = 1)
}
\arguments{
\item{model}{Bayesian model object}

\item{scale}{"response" or "lin_pred"}

\item{model_name}{provides a name for the model}

\item{newdata}{new data to predict from}

\item{thin}{thinning factor}
}
\value{
tbl_postpred object with MCMC draws

chains are stored in a
long table with the following columns:

chain  iter   parameter   value  type  order
(fctr) (int)  (chr)       (dbl)  (chr) (int)
}
\description{
MCMC predicted values are  extracted from a Bayesian (regression) object
and returned as a tbl_post_pred object
}
\author{
Martin Schmettow
}
