% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup_helpers.R, R/posterior_extraction.R
\name{print.tbl_post}
\alias{print.tbl_post}
\alias{knit_print.tbl_post}
\alias{posterior}
\alias{tbl_post}
\alias{tbl_post.data.frame}
\alias{tbl_post.brmsfit}
\alias{tbl_post.stanreg}
\title{posterior extraction}
\usage{
\method{print}{tbl_post}(tbl_post, ...)

\method{knit_print}{tbl_post}(x, ...)

posterior(model, shape, ...)

tbl_post(model, ...)

\method{tbl_post}{data.frame}(df, ...)

\method{tbl_post}{brmsfit}(model, ...)

\method{tbl_post}{stanreg}(model, ...)
}
\arguments{
\item{model}{Bayesian model object}

\item{shape}{return tbl_post in long shape or tbl_df in wide shape}

\item{thin}{thinning factor}

\item{type}{select parameter types to store ("fixef", "grpef", "ranef")}

\item{model_name}{provides a name for the model}
}
\value{
tbl_post object with MCMC chain in long format or tbl.df in wide shape

The MCMC chains are extracted from the model and stored in a
common format. Different to most internal representations, and
mcmc objects from package coda, the chains are stored in a
long table with the following columns:

chain  iter   parameter   value  type  order
(fctr) (int)  (chr)       (dbl)  (chr) (int)
}
\description{
MCMC chains are  extracted from a Bayesian (regression) object
and returned as a posterior object, which is in long format
(chain, iter, parameter, value, type, order). Parameters are classified
as fixef, ranef, grpef and named after a common scheme.
}
\author{
Martin Schmettow
}
