% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_extraction.R
\name{predict.tbl_post_pred}
\alias{predict.tbl_post_pred}
\alias{predict}
\alias{predict.brmsfit}
\alias{predict.stanreg}
\title{Posterior predictions}
\usage{
\method{predict}{tbl_post_pred}(object, model = unique(object$model),
  scale = c("resp"), center = median, interval = 0.95, ...)

predict(object, estimate = median, ...)

\method{predict}{brmsfit}(object, center = median, ...)

\method{predict}{stanreg}(object, center = median, ...)
}
\arguments{
\item{object}{tbl_post_pred object holding the predictive posterior in long format}

\item{model}{model}

\item{scale}{linpred or resp}

\item{center}{function for computing the center estimate (median)}

\item{interval}{credibility interval: .95}

\item{...}{ignored}
}
\value{
coefficient table with center and interval estimates per obs

The standard center function is the posterior median
}
\description{
summary table of predicted values from predictive posterior
}
\author{
Martin Schmettow
}
