% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_manip.R
\name{re_scores}
\alias{re_scores}
\title{absolute random effects scores}
\usage{
re_scores(tbl_post)
}
\arguments{
\item{tbl_post}{posterior (tbl_post)}

\item{type}{set type of coefficient (ranef)}
}
\value{
tbl_post

Random coefficients are estimated as differences to the population level coefficient.
In some situations one needs the total (or absolute) score.

Example: It is typical for psychometric settings to think of item and
person scores. An average person has an IQ of 100,
but would have a zero in terms of random effects.
}
\description{
posterior distribution of absolute random effects
}
\author{
Martin Schmettow
}
