% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_manip.R
\name{re_scores}
\alias{re_scores}
\title{total random effects scores}
\usage{
re_scores(tbl_post, fixef, ranef)
}
\arguments{
\item{tbl_post}{posterior (tbl_post)}

\item{re_factor}{random factor (char)}
}
\value{
tbl_post

Random coefficients are created as differences to the group mean.
In some situations one needs the total (or absolute) score.

Example: It is typical for psychometric settings to think of item and
person scores. An average person has an IQ of 100,
but would have a zero in terms of random effects.
}
\description{
sums over random effects and their respective population-level effects
}
\author{
Martin Schmettow
}
