% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_extensions.R
\name{rescale_unit}
\alias{rescale_unit}
\alias{rescale_zero_one}
\title{rescaling to unit interval}
\usage{
rescale_unit(
  x,
  lower = min(x, na.rm = T),
  upper = max(x, na.rm = T),
  scale = 1
)

rescale_zero_one(
  x,
  lower = min(x, na.rm = T),
  upper = max(x, na.rm = T),
  scale = 1
)
}
\arguments{
\item{x}{numerical vector}

\item{lower}{lowest possible value}

\item{upper}{highest possible value}

\item{scale}{rescale factor}
}
\value{
numerical vector
}
\description{
scales a variable to the interval [0;1]
}
\author{
Martin Schmettow
}
